Replace all State with Refs in custom hooks useTuringMachine and useTape
So this is... not such a good idea. Now if the state of the turing machine changes it doesn't cause a re-render which is generally not what you want.
However because we always modify some other States along with the turing machine, the re-renders happen anyway hence it works.

I needed to do it this way so I could call the methods from the hooks multiple times in the same state snapshot, and have these methods pull the latest state.